requirements.txt
# Core Backend
fastapi
uvicorn[standard]
jinja2
python-multipart

# AI Agent & Browser Automation
browser-use
langchain-google-genai
playwright
python-dotenv
